refactor: break circular dependencies to fix test-unit (#18237)
