+++ packages/vite/src/node/constants.ts
@@ -173,3 +173,9 @@ export const DEFAULT_PREVIEW_PORT = 4173
 export const DEFAULT_ASSETS_INLINE_LIMIT = 4096
 
 export const METADATA_FILENAME = '_metadata.json'
+
+export const ERR_OPTIMIZE_DEPS_PROCESSING_ERROR =
+  'ERR_OPTIMIZE_DEPS_PROCESSING_ERROR'
+export const ERR_OUTDATED_OPTIMIZED_DEP = 'ERR_OUTDATED_OPTIMIZED_DEP'
+export const ERR_FILE_NOT_FOUND_IN_OPTIMIZED_DEP_DIR =
+  'ERR_FILE_NOT_FOUND_IN_OPTIMIZED_DEP_DIR'
