+++ packages/vite/src/node/server/middlewares/transform.ts
@@ -21,17 +21,18 @@ import { send } from '../send'
 import { ERR_LOAD_URL, transformRequest } from '../transformRequest'
 import { applySourcemapIgnoreList } from '../sourcemap'
 import { isHTMLProxy } from '../../plugins/html'
-import { DEP_VERSION_RE, FS_PREFIX } from '../../constants'
+import {
+  DEP_VERSION_RE,
+  ERR_FILE_NOT_FOUND_IN_OPTIMIZED_DEP_DIR,
+  ERR_OPTIMIZE_DEPS_PROCESSING_ERROR,
+  ERR_OUTDATED_OPTIMIZED_DEP,
+  FS_PREFIX,
+} from '../../constants'
 import {
   isCSSRequest,
   isDirectCSSRequest,
   isDirectRequest,
 } from '../../plugins/css'
-import {
-  ERR_FILE_NOT_FOUND_IN_OPTIMIZED_DEP_DIR,
-  ERR_OPTIMIZE_DEPS_PROCESSING_ERROR,
-  ERR_OUTDATED_OPTIMIZED_DEP,
-} from '../../plugins/optimizedDeps'
 import { ERR_CLOSED_SERVER } from '../pluginContainer'
 import { cleanUrl, unwrapId, withTrailingSlash } from '../../../shared/utils'
 import { NULL_BYTE_PLACEHOLDER } from '../../../shared/constants'
